move child components to parents component foldeR
