+++ src/app/facility/utility-data/energy-consumption/energy-consumption.module.ts
@@ -2,8 +2,8 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { EditMeterComponent } from './energy-source/edit-meter/edit-meter.component';
 import { EnergySourceComponent } from './energy-source/energy-source.component';
-import { EditElectricityBillComponent } from './utility-meter-data/edit-electricity-bill/edit-electricity-bill.component';
-import { EditUtilityBillComponent } from './utility-meter-data/edit-utility-bill/edit-utility-bill.component';
+import { EditElectricityBillComponent } from './utility-meter-data/edit-bill/edit-electricity-bill/edit-electricity-bill.component';
+import { EditUtilityBillComponent } from './utility-meter-data/edit-bill/edit-utility-bill/edit-utility-bill.component';
 import { ElectricityDataTableComponent } from './utility-meter-data/utility-meter-data-table/electricity-data-table/electricity-data-table.component';
 import { GeneralUtilityDataTableComponent } from './utility-meter-data/utility-meter-data-table/general-utility-data-table/general-utility-data-table.component';
 import { UtilityMeterDataFilterComponent } from './utility-meter-data/utility-meter-data-filter/utility-meter-data-filter.component';
@@ -18,7 +18,7 @@ import { UtilityMetersTableComponent } from './energy-source/utility-meters-tabl
 import { UtilityMeterDataTableComponent } from './utility-meter-data/utility-meter-data-table/utility-meter-data-table.component';
 import { EditBillComponent } from './utility-meter-data/edit-bill/edit-bill.component';
 import { TableItemsDropdownModule } from 'src/app/shared/table-items-dropdown/table-items-dropdown.module';
-import { EditVehicleMeterBillComponent } from './utility-meter-data/edit-vehicle-meter-bill/edit-vehicle-meter-bill.component';
+import { EditVehicleMeterBillComponent } from './utility-meter-data/edit-bill/edit-vehicle-meter-bill/edit-vehicle-meter-bill.component';
 import { VehicleDataTableComponent } from './utility-meter-data/utility-meter-data-table/vehicle-data-table/vehicle-data-table.component';
 @NgModule({
   declarations: [
