+++ src/app/facility/utility-data/energy-consumption/utility-meter-data/edit-bill/edit-electricity-bill/edit-electricity-bill.component.ts
@@ -4,7 +4,7 @@ import { Subscription } from 'rxjs';
 import { UtilityMeterDatadbService } from 'src/app/indexedDB/utilityMeterData-db.service';
 import { AdditionalChargesFilters, DetailedChargesFilters, ElectricityDataFilters } from 'src/app/models/meterDataFilter';
 import { IdbCustomFuel, IdbFacility, IdbUtilityMeter, IdbUtilityMeterData } from 'src/app/models/idb';
-import { UtilityMeterDataService } from '../utility-meter-data.service';
+import { UtilityMeterDataService } from '../../utility-meter-data.service';
 import { EmissionsResults } from 'src/app/models/eGridEmissions';
 import { getEmissions } from 'src/app/calculations/emissions-calculations/emissions';
 import { EGridService } from 'src/app/shared/helper-services/e-grid.service';
